requirements/base.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers
django-redis==5.4.0
pycryptodome==3.20.0
requests==2.32.3